tests/wpt/tests/html/semantics/scripting-1/the-script-element/module/single-evaluation-2.html

<!DOCTYPE html>
<title>Single evaluation, 2</title>

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
    window.log = [];

    const test_load = async_test(
        "Test that a module is evaluated only once, and that 'this' is " +
        "undefined (because of strict mode).");
    window.addEventListener("load", test_load.step_func_done(ev => {
      assert_array_equals(log, [undefined, "this-nested"]);
    }));
</script>
<script type="module" src="this-nested.js"></script>
<script type="module" src="this-nested.js"></script>
<script type="module" src="this.js"></script>
<script type="module" src="this-nested.js"></script>
<script type="module" src="this.js"></script>
